refactor: 도메인명 변경
