+++ back/src/main/java/com/ll/goohaeyou/domain/image/service/JobPostImageService.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.domain.fileupload.service;
+package com.ll.goohaeyou.domain.image.service;
 
 import com.ll.goohaeyou.domain.jobPost.jobPost.entity.JobPostDetail;
 import com.ll.goohaeyou.domain.jobPost.jobPost.entity.JobPostImage;
