+++ back/src/main/java/com/ll/goohaeyou/domain/image/service/S3ImageService.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.domain.fileupload.service;
+package com.ll.goohaeyou.domain.image.service;
 
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.SdkClientException;
